Add ChoiceTransition.

+++ src/transition.h
@@ -66,7 +66,16 @@ struct OkGuard {
 };
 
 namespace impl {
-template<uint8_t Trigger, typename To, typename From, typename CreationPolicy, typename Guard, typename Action, bool IsExitingTransition, bool IsEnteringTransition>
+template<
+  uint8_t Trigger,
+  typename To,
+  typename To2,
+  typename From,
+  typename CreationPolicy,
+  typename Guard,
+  typename Action,
+  bool IsExitingTransition,
+  bool IsEnteringTransition>
 struct TransitionBase {
   enum { N = Trigger };
   enum { E = IsEnteringTransition };
@@ -77,13 +86,14 @@ struct TransitionBase {
 
   DispatchResult<StateType> dispatch(StateType* activeState) {
     typedef typename To::CreatorType ToFactory;
+    typedef typename To2::CreatorType To2Factory;
     typedef typename From::CreatorType FromFactory;
-    To* toState = ToFactory::create();
     From* fromState = FromFactory::create();
 
     // Initial transition
     if (!is_same<EmptyState<StateType>, To>().value && is_same<EmptyState<StateType>, From>().value) {
       Action().perform(activeState);
+      To* toState = ToFactory::create();
       bool cosumedBySubstate = toState->template _entry<N>();
       if (!cosumedBySubstate) {
         toState->template _doit<N>();
@@ -97,7 +107,7 @@ struct TransitionBase {
     // End transition; to any state <- from AnyState
     if (is_same<From, AnyState<StateType>>().value) {
 
-      // Delete toState and fromState not needed; both are "null".
+      // Delete fromState not needed; it is "null".
 
       if (Guard().check(activeState)) {
         // TODO: "exit" of AnyState is called, not from the activeState object. Polymorphism is required.
@@ -108,6 +118,7 @@ struct TransitionBase {
         typedef typename CreationPolicy::CreatorType Creator;
         typedef typename CreationPolicy::ObjectType Object;
         Creator::destroy(static_cast<Object*>(activeState));
+        To* toState = ToFactory::create();
         return DispatchResult<StateType>(true, toState);
       }
       return DispatchResult<StateType>(false, activeState);
@@ -118,6 +129,7 @@ struct TransitionBase {
 
       // Entering substate transition
       if (E && activeState == 0) {
+        To* toState = ToFactory::create();
         bool cosumedBySubstate = toState->template _entry<N>();
         if (!cosumedBySubstate) {
           toState->template _doit<N>();
@@ -128,23 +140,61 @@ struct TransitionBase {
         return DispatchResult<StateType>(true, toState);
       }
 
-      ToFactory::destroy(toState);
       FromFactory::destroy(fromState);
       return DispatchResult<StateType>(false, activeState);
     }
-    FromFactory::destroy(fromState);
 
     Action().perform(activeState);
 
+    // Choice transition
+    if (!is_same<To2, EmptyState<typename CreationPolicy::ObjectType>>().value) {
+
+      if (Guard().check(activeState)) {
+        To* toState = ToFactory::create();
+
+        // Self transition
+        if (!fromState->equals(*toState)) {
+          static_cast<From*>(activeState)->_exit();
+          FromFactory::destroy(static_cast<From*>(activeState));
+        }
+        FromFactory::destroy(fromState);
+        bool cosumedBySubstate = toState->template _entry<N>();
+        if (!cosumedBySubstate) {
+          toState->template _doit<N>();
+        }
+        return DispatchResult<StateType>(true, toState, X);
+      }
+      else {
+        To2* toState = To2Factory::create();
+
+        // Self transition
+        if (fromState->equals(*toState)) {
+          toState->template _doit<N>();
+          return DispatchResult<StateType>(true, toState, X);
+        }
+
+        FromFactory::destroy(static_cast<From*>(activeState));
+        static_cast<From*>(activeState)->_exit();
+        FromFactory::destroy(fromState);
+        bool cosumedBySubstate = toState->template _entry<N>();
+        if (!cosumedBySubstate) {
+          toState->template _doit<N>();
+        }
+        return DispatchResult<StateType>(true, toState, X);
+      }
+    }
+
+    FromFactory::destroy(fromState);
+
     if (!Guard().check(activeState)) {
-      ToFactory::destroy(toState);
       return DispatchResult<StateType>(false, activeState);
     }
+
+    
 
     // Self transition
     if (is_same<To, From>().value) {
       auto state = static_cast<To*>(activeState)->template _doit<N>();
-      ToFactory::destroy(toState);
       return DispatchResult<StateType>(true, state != 0 ? state : activeState);
     }
 
@@ -152,9 +202,11 @@ struct TransitionBase {
 
     if (X) {
       FromFactory::destroy(static_cast<From*>(activeState));
+      To* toState = ToFactory::create();
       return DispatchResult<StateType>(true, toState, X);
     }
 
+    To* toState = ToFactory::create();
     bool cosumedBySubstate = toState->template _entry<N>();
     if (!cosumedBySubstate) {
       toState->template _doit<N>();
@@ -166,11 +218,11 @@ struct TransitionBase {
 }
 
 template<typename To, typename CreationPolicy, typename Action>
-struct InitialTransition : impl::TransitionBase<0, To, EmptyState<typename CreationPolicy::ObjectType>, CreationPolicy, OkGuard, Action, false, true> {
+struct InitialTransition : impl::TransitionBase<0, To, EmptyState<typename CreationPolicy::ObjectType>, EmptyState<typename CreationPolicy::ObjectType>, CreationPolicy, OkGuard, Action, false, true> {
 };
 
 template<uint8_t Trigger, typename From, typename CreationPolicy, typename Guard, typename Action>
-struct FinalTransition : impl::TransitionBase<Trigger, EmptyState<typename CreationPolicy::ObjectType>, From, CreationPolicy, Guard, Action, false, false> {
+struct FinalTransition : impl::TransitionBase<Trigger, EmptyState<typename CreationPolicy::ObjectType>, EmptyState<typename CreationPolicy::ObjectType>, From, CreationPolicy, Guard, Action, false, false> {
   FinalTransition() {
     // Make sure the user defines a guard for the final transition. This is not UML compliant.
     CompileTimeError < !is_same<Guard, OkGuard>().value > ();
@@ -178,7 +230,7 @@ struct FinalTransition : impl::TransitionBase<Trigger, EmptyState<typename Creat
 };
 
 template<typename CreationPolicy, typename Guard, typename Action>
-struct EndTransition : impl::TransitionBase<0, EmptyState<typename CreationPolicy::ObjectType>, AnyState<typename CreationPolicy::ObjectType>, CreationPolicy, Guard, Action, false, false> {
+struct EndTransition : impl::TransitionBase<0, EmptyState<typename CreationPolicy::ObjectType>, EmptyState<typename CreationPolicy::ObjectType>, AnyState<typename CreationPolicy::ObjectType>, CreationPolicy, Guard, Action, false, false> {
   EndTransition() {
     // Final transition without guard does not make sense; the state machine would immediately go to the final state.
     CompileTimeError < !is_same<Guard, OkGuard>().value > ();
@@ -186,24 +238,30 @@ struct EndTransition : impl::TransitionBase<0, EmptyState<typename CreationPolic
 };
 
 template<typename CreationPolicy>
-using NullEndTransition = impl::TransitionBase<0, EmptyState<typename CreationPolicy::ObjectType>, AnyState<typename CreationPolicy::ObjectType>, CreationPolicy, OkGuard, EmptyAction, false, false>;
+using NullEndTransition = impl::TransitionBase<0, EmptyState<typename CreationPolicy::ObjectType>, EmptyState<typename CreationPolicy::ObjectType>, AnyState<typename CreationPolicy::ObjectType>, CreationPolicy, OkGuard, EmptyAction, false, false>;
 
 template<uint8_t Trigger, typename Me, typename CreationPolicy, typename Guard, typename Action>
-using SelfTransition = impl::TransitionBase<Trigger, Me, Me, CreationPolicy, Guard, Action, false, false>;
+using SelfTransition = impl::TransitionBase<Trigger, Me, EmptyState<typename CreationPolicy::ObjectType>, Me, CreationPolicy, Guard, Action, false, false>;
 
 template<uint8_t Trigger, typename Me, typename CreationPolicy>
-using Declaration = impl::TransitionBase<Trigger, Me, Me, CreationPolicy, OkGuard, EmptyAction, false, false>;
+using Declaration = impl::TransitionBase<Trigger, Me, EmptyState<typename CreationPolicy::ObjectType>, Me, CreationPolicy, OkGuard, EmptyAction, false, false>;
 
 template<uint8_t Trigger, typename To, typename Me, typename CreationPolicy, typename Action = EmptyAction>
-using ExitDeclaration = impl::TransitionBase<Trigger, To, Me, CreationPolicy, OkGuard, Action, true, false>;
+using ExitDeclaration = impl::TransitionBase<Trigger, To, EmptyState<typename CreationPolicy::ObjectType>, Me, CreationPolicy, OkGuard, Action, true, false>;
 
 template<uint8_t Trigger, typename To, typename CreationPolicy, typename Action>
-using EntryDeclaration = impl::TransitionBase<Trigger, To, To, CreationPolicy, OkGuard, Action, false, true>;
+using EntryDeclaration = impl::TransitionBase<Trigger, To, EmptyState<typename CreationPolicy::ObjectType>, To, CreationPolicy, OkGuard, Action, false, true>;
 
 template<uint8_t Trigger, typename To, typename From, typename CreationPolicy, typename Guard, typename Action>
-using ExitTransition = impl::TransitionBase<Trigger, To, From, CreationPolicy, Guard, Action, true, false>;
+using ExitTransition = impl::TransitionBase<Trigger, To, EmptyState<typename CreationPolicy::ObjectType>, From, CreationPolicy, Guard, Action, true, false>;
 
 template<uint8_t Trigger, typename To, typename From, typename CreationPolicy, typename Guard, typename Action>
-using Transition = impl::TransitionBase<Trigger, To, From, CreationPolicy, Guard, Action, false, false>;
+using Transition = impl::TransitionBase<Trigger, To, EmptyState<typename CreationPolicy::ObjectType>, From, CreationPolicy, Guard, Action, false, false>;
+
+template<uint8_t Trigger, typename To_true, typename To_false, typename From, typename CreationPolicy, typename Guard, typename Action>
+struct ChoiceTransition : impl::TransitionBase<Trigger, To_true, To_false, From, CreationPolicy, Guard, Action, false, false> {
+  // Choice without guard does not make sense; the choice is either to go the the state To_true or To_false.
+  ChoiceTransition() { CompileTimeError < !is_same<Guard, OkGuard>().value >(); }
+};
 
 }
